# infra/rest/swagger/go/business_partner.go
/*
 * Dhanraj Hardware Stores
 *
 * Dhanraj Hardware Stores APIs
 *
 * API version: 1.0.0
 * Contact: viveks.singh@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type BusinessPartner struct {

	// Unique ID of the Business Partner
	Id int32 `json:"id,omitempty"`

	Basicinfo *BusinessPartnerBasicinfo `json:"basicinfo,omitempty"`

	// Introduced By
	Introducedby string `json:"introducedby,omitempty"`
}