# infra/rest/swagger/go/business_partners.go
/*
 * Dhanraj Hardware Stores
 *
 * Dhanraj Hardware Stores APIs
 *
 * API version: 1.0.0
 * Contact: viveks.singh@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type BusinessPartners struct {

	Bplist []BusinessPartner `json:"bplist,omitempty"`
}