# infra/rest/swagger/go/error_model.go
/*
 * Dhanraj Hardware Stores
 *
 * Dhanraj Hardware Stores APIs
 *
 * API version: 1.0.0
 * Contact: viveks.singh@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ErrorModel struct {

	// 1000 - No Error. 1001 - Validation Error. 1101 - Assignment Error. 1301 - Database Error. 
	Code int32 `json:"code"`

	// Describing the reason of failure
	Message string `json:"message"`
}