# infra/rest/swagger/go/sale_order.go
/*
 * Dhanraj Hardware Stores
 *
 * Dhanraj Hardware Stores APIs
 *
 * API version: 1.0.0
 * Contact: viveks.singh@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

import (
	"time"
)

type SaleOrder struct {

	Id int64 `json:"id,omitempty"`

	Bpid int64 `json:"bpid,omitempty"`

	OrderDate time.Time `json:"order_date,omitempty"`

	ShipDate time.Time `json:"ship_date,omitempty"`

	// Payment Status
	DeliveryStatus string `json:"delivery_status,omitempty"`

	Amount int32 `json:"amount,omitempty"`

	// Payment Status
	PaymentStatus string `json:"payment_status,omitempty"`

	// Who packed the order
	Packedby string `json:"packedby,omitempty"`

	// Who checked the order
	Checkedby string `json:"checkedby,omitempty"`

	// transport name
	Transport string `json:"transport,omitempty"`

	// Any free text about this order
	Remark string `json:"remark,omitempty"`
}