# infra/rest/swagger/go/sale_orders.go
/*
 * Dhanraj Hardware Stores
 *
 * Dhanraj Hardware Stores APIs
 *
 * API version: 1.0.0
 * Contact: viveks.singh@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type SaleOrders struct {

	SaleOrderList []SaleOrder `json:"sale_order_list,omitempty"`
}